+++ app/Services/MethodService.php
@@ -24,7 +24,7 @@ class MethodService extends BaseService
     public function searchPaginate($search = null)
     {
         $user = Auth::user();
-        if (! $user->can('kiims.view')) {
+        if (! $user->can('landing.view')) {
             throw new GeneralException(__('You do not have access to do that.'));
         }
 
@@ -40,10 +40,10 @@ class MethodService extends BaseService
 
     public function datatables($search = '')
     {
-//        $user = Auth::user();
-//        if (! $user->can('kiims.view')) {
-//            throw new GeneralException(__('You do not have access to do that.'));
-//        }
+       $user = Auth::user();
+       if (! $user->can('landing.view')) {
+           throw new GeneralException(__('You do not have access to do that.'));
+       }
         $query = $this->model->query();
 
         if (! empty($search)) {
@@ -55,10 +55,10 @@ class MethodService extends BaseService
 
     public function store(array $data = []): Method
     {
-        // $user = Auth::user();
-        // if (! $user->can('kiims.create')) {
-        //     throw new GeneralException(__('You do not have access to do that.'));
-        // }
+        $user = Auth::user();
+        if (! $user->can('landing.create')) {
+            throw new GeneralException(__('You do not have access to do that.'));
+        }
 
         return  $this->model::create($data);
     }
@@ -70,20 +70,20 @@ class MethodService extends BaseService
      */
     public function update(Method $item, array $data = []): bool
     {
-        // $user = Auth::user();
-        // if (! $user->can('kiims.edit')) {
-        //     throw new GeneralException(__('You do not have access to do that.'));
-        // }
+        $user = Auth::user();
+        if (! $user->can('landing.edit')) {
+            throw new GeneralException(__('You do not have access to do that.'));
+        }
 
         return $item->update($data);
     }
 
     public function delete(Method $item, array $data = []): bool
     {
-        // $user = Auth::user();
-        // if (! $user->can('kiims.edit')) {
-        //     throw new GeneralException(__('You do not have access to do that.'));
-        // }
+        $user = Auth::user();
+        if (! $user->can('landing.edit')) {
+            throw new GeneralException(__('You do not have access to do that.'));
+        }
 
         return $item->delete($data);
     }
