+++ app/Services/SpeciesService.php
@@ -24,10 +24,10 @@ class SpeciesService extends BaseService
 
     public function searchPaginate($search = null)
     {
-        // $user = Auth::user();
-        // if (! $user->can('kiims.view')) {
-        //     throw new GeneralException(__('You do not have access to do that.'));
-        // }
+        $user = Auth::user();
+        if (! $user->can('landing.view')) {
+            throw new GeneralException(__('You do not have access to do that.'));
+        }
 
         $query = $this->model->query();
         if (! empty($search)) {
@@ -39,10 +39,10 @@ class SpeciesService extends BaseService
 
     public function datatables($search = '')
     {
-        // $user = Auth::user();
-        // if (! $user->can('kiims.view')) {
-        //     throw new GeneralException(__('You do not have access to do that.'));
-        // }
+        $user = Auth::user();
+        if (! $user->can('landing.view')) {
+            throw new GeneralException(__('You do not have access to do that.'));
+        }
         $query = $this->model->query();
 
         if (! empty($search)) {
@@ -60,10 +60,10 @@ class SpeciesService extends BaseService
      */
     public function store(array $data = []): Species
     {
-        // $user = Auth::user();
-        // if (! $user->can('kiims.create')) {
-        //     throw new GeneralException(__('Not Authorised'));
-        // }
+        $user = Auth::user();
+        if (! $user->can('landing.create')) {
+            throw new GeneralException(__('Not Authorised'));
+        }
 
         return  $this->model::create($data);
     }
@@ -76,20 +76,20 @@ class SpeciesService extends BaseService
      */
     public function update(Species $item, array $data = []): bool
     {
-        // $user = Auth::user();
-        // if (! $user->can('kiims.edit')) {
-        //     throw new GeneralException(__('You do not have access to do that.'));
-        // }
+        $user = Auth::user();
+        if (! $user->can('landing.edit')) {
+            throw new GeneralException(__('You do not have access to do that.'));
+        }
 
         return $item->update($data);
     }
 
     public function delete(Species $item, array $data = []): bool
     {
-        // $user = Auth::user();
-        // if (! $user->can('kiims.edit')) {
-        //     throw new GeneralException(__('You do not have access to do that.'));
-        // }
+        $user = Auth::user();
+        if (! $user->can('landing.edit')) {
+            throw new GeneralException(__('You do not have access to do that.'));
+        }
 
         return $item->delete($data);
     }
